implemented ArrayModifierAdapter.InMap

+++ modifier/modifier_suite_test.go
@@ -248,7 +248,7 @@ var _ = Describe("Modifier", func() {
 				Context("When I read the head of the array", func() {
 					var r io.Reader
 					BeforeEach(func() {
-						er := m.EntityReaderFor(dbpath.Path{"l1", 0})
+						er := m.EntityReaderFor(dbpath.Path{"l1", uint64(0)})
 						r = er.Data()
 					})
 					Context("When I read the value", func() {
@@ -277,7 +277,7 @@ var _ = Describe("Modifier", func() {
 					Context("When I read the head of the array", func() {
 						var r io.Reader
 						BeforeEach(func() {
-							er := m.EntityReaderFor(dbpath.Path{"l1", 0})
+							er := m.EntityReaderFor(dbpath.Path{"l1", uint64(0)})
 							r = er.Data()
 						})
 						Context("When I read the value", func() {
@@ -296,7 +296,7 @@ var _ = Describe("Modifier", func() {
 					Context("When I read the second value of the array", func() {
 						var r io.Reader
 						BeforeEach(func() {
-							er := m.EntityReaderFor(dbpath.Path{"l1", 1})
+							er := m.EntityReaderFor(dbpath.Path{"l1", uint64(1)})
 							r = er.Data()
 						})
 						Context("When I read the value", func() {
